refactor: move services into model
