+++ src/entities/Cart/model/services/addToCart.ts
@@ -1,7 +1,7 @@
 import { createAsyncThunk } from '@reduxjs/toolkit';
 import axios from 'axios';
 import { BaseTokenError, ErrorWithResponse } from 'shared/types/errorResponseTypes';
-import { ActionCartProps, Cart } from '../model/types/cartTypes';
+import { ActionCartProps, Cart } from '../types/cartTypes';
 
 const PROJECT_KEY = process.env.PROJECT_KEY;
 const API_URL = process.env.API_URL;
@@ -27,7 +27,13 @@ export const addToCart = createAsyncThunk('cart/addToCart', async (props: AddToC
     const headers = { Authorization: `Bearer ${token}`, 'Content-Type': 'application/json' };
     const res = await axios.post<Cart>(`${API_URL}${PROJECT_KEY}/carts/${cartId}`, body, { headers });
 
-    const success: Cart = { id: res.data.id, version: res.data.version, lineItems: res.data.lineItems };
+    const success: Cart = {
+      id: res.data.id,
+      version: res.data.version,
+      lineItems: res.data.lineItems,
+      totalPrice: res.data.totalPrice,
+    };
+
     return success;
   } catch (error) {
     let errorMsg = 'error';
