+++ src/entities/Cart/model/services/clearRemoteCart.ts
@@ -1,37 +1,45 @@
 import { createAsyncThunk } from '@reduxjs/toolkit';
 import axios from 'axios';
 import { ErrorWithResponse, BaseTokenError } from 'shared/types';
-import { ActionCartProps, Cart } from '../model/types/cartTypes';
+import { Cart } from '../types/cartTypes';
 
 const PROJECT_KEY = process.env.PROJECT_KEY;
 const API_URL = process.env.API_URL;
 
-interface RemoveFromCartProps extends ActionCartProps {
-  lineItemId: string;
+interface RemoveFromCartProps {
+  token: string;
+  version: number;
+  cartId: string;
+  lineItemId: string[];
 }
 
-export const removeFromCart = createAsyncThunk(
-  'cart/removeFromCart',
+export const clearRemoteCart = createAsyncThunk(
+  'cart/clearRemoteCart',
   async (props: RemoveFromCartProps, { rejectWithValue }) => {
     try {
-      const { token, version, cartId, count, lineItemId } = props;
+      const { token, version, lineItemId, cartId } = props;
+
+      const actions = lineItemId.map((item) => {
+        return {
+          action: 'removeLineItem',
+          lineItemId: item,
+        };
+      });
 
       const body = {
         version,
-        actions: [
-          {
-            action: 'removeLineItem',
-            lineItemId,
-            quantity: count || 1,
-          },
-        ],
+        actions,
       };
 
       const headers = { Authorization: `Bearer ${token}`, 'Content-Type': 'application/json' };
       const res = await axios.post<Cart>(`${API_URL}${PROJECT_KEY}/carts/${cartId}`, body, { headers });
 
-      const success: Cart = { id: res.data.id, version: res.data.version, lineItems: res.data.lineItems };
-
+      const success: Cart = {
+        id: res.data.id,
+        version: res.data.version,
+        lineItems: res.data.lineItems,
+        totalPrice: res.data.totalPrice,
+      };
       return success;
     } catch (error) {
       let errorMsg = 'error';
