+++ src/entities/Cart/model/services/removeFromCart.ts
@@ -1,40 +1,42 @@
 import { createAsyncThunk } from '@reduxjs/toolkit';
 import axios from 'axios';
 import { ErrorWithResponse, BaseTokenError } from 'shared/types';
-import { Cart } from '../model/types/cartTypes';
+import { ActionCartProps, Cart } from '../types/cartTypes';
 
 const PROJECT_KEY = process.env.PROJECT_KEY;
 const API_URL = process.env.API_URL;
 
-interface RemoveFromCartProps {
-  token: string;
-  version: number;
-  cartId: string;
-  lineItemId: string[];
+interface RemoveFromCartProps extends ActionCartProps {
+  lineItemId: string;
 }
 
-export const clearRemoteCart = createAsyncThunk(
-  'cart/clearRemoteCart',
+export const removeFromCart = createAsyncThunk(
+  'cart/removeFromCart',
   async (props: RemoveFromCartProps, { rejectWithValue }) => {
     try {
-      const { token, version, lineItemId, cartId } = props;
-
-      const actions = lineItemId.map((item) => {
-        return {
-          action: 'removeLineItem',
-          lineItemId: item,
-        };
-      });
+      const { token, version, cartId, count, lineItemId } = props;
 
       const body = {
         version,
-        actions,
+        actions: [
+          {
+            action: 'removeLineItem',
+            lineItemId,
+            quantity: count || 1,
+          },
+        ],
       };
 
       const headers = { Authorization: `Bearer ${token}`, 'Content-Type': 'application/json' };
       const res = await axios.post<Cart>(`${API_URL}${PROJECT_KEY}/carts/${cartId}`, body, { headers });
 
-      const success: Cart = { id: res.data.id, version: res.data.version, lineItems: res.data.lineItems };
+      const success: Cart = {
+        id: res.data.id,
+        version: res.data.version,
+        lineItems: res.data.lineItems,
+        totalPrice: res.data.totalPrice,
+      };
+
       return success;
     } catch (error) {
       let errorMsg = 'error';
